requirements.txt
gym-retro==0.8.0
gym==0.21.0
argparse==1.4.0
numpy==1.18.5
tqdm==4.59.0
futures==3.1.1
